Change messages to use the correct bit-sized containers. Provide Unsigned utility class to convert unsigned values to signed values in larger containers. Add support for array sequences and change all primitive type sequences to use arrays. Refactor MessageContext, FieldType, and Field to make it easier to make these changes.

+++ rosjava/src/test/java/org/ros/internal/message/MessageSerializationTest.java
@@ -92,7 +92,7 @@ public class MessageSerializationTest {
   @Test
   public void testUint8() {
     std_msgs.UInt8 message = defaultMessageFactory.newFromType(std_msgs.UInt8._TYPE);
-    message.setData((short) 42);
+    message.setData((byte) 42);
     checkSerializeAndDeserialize(message);
   }
